requirements.txt
h5py==2.7.1
matplotlib==2.2.2
Pillow==5.1.0
scikit-learn==0.19.1
tensorflow==1.10.0
tensorflow-gpu==1.10.0